Revert "Pierre's work"
